+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### next release (8.3.3)
 
+- Allow a story to use iframe tag
 - [The next improvement]
 
 #### 8.3.2 - 2023-08-11